requirements.txt
asgiref==3.5.0
astroid==2.9.3
autopep8==1.6.0
backports.zoneinfo==0.2.1
Django==4.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-action-serializer==1.0.5
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
platformdirs==2.4.1
pycodestyle==2.8.0
PyJWT==2.3.0
pylint==2.12.2
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.0.1
wrapt==1.13.3
